requirements.txt
streamlit
pandas
plotly-express
matplotlib
seaborn
joblib
numpy
scikit-learn
altair
streamlit-option-menu
on-Hover-tabs